Show live team and inventory counts on home
